pyproject.toml
[build-system]
requires = ["setuptools>=42", "setuptools_scm[toml]>=6.2"]   # PEP 508 specifications.
build-backend = "setuptools.build_meta"

[project]
name = "resumy"
authors = [
    {name = "Alex Laurent", email="alex@i42.sh"},
]
description = "Resume builder"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["resume", "cv"]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3.8",
]
dependencies = [
    "PyYAML==6.0",
    "Jinja2>=3.1.2",
    "weasyprint>=56.1"
]
dynamic = ["version"]

[project.urls]
homepage = "https://github.com/alexlren/resumy"

[project.scripts]
resumy = "resumy.resumy:main"

[tool.setuptools_scm]
